refactor(data): merge yearly_files.py into input_resolution.py
The {year} file-pattern expansion was a separate import-free leaf only so
tools/benchmark could load it by file path (jcm-free) and share it with the
resolver. That resolver is jcm.data.input_resolution — itself import-free at
module top — so the leaf is redundant: fold expand_yearly_files into it as one
jax-free engine module. jcm.forcing keeps re-exporting expand_yearly_files (now
sourced from input_resolution), so the runner and tests are unchanged.

benchmark's file-path loader now loads input_resolution.py; because that module
defines dataclasses (whose deferred-annotation resolution consults
sys.modules), the loader registers the module in sys.modules before exec — the
fix the merge required. Verified: the loader stays jcm-free (the module's jcm
imports are lazy, never hit by expand_yearly_files).
